app/scripts/qaqc/requirements.txt
requests>=2.26
rasterio
geopandas>=0.10
jupyterlab>=3.2.1
python-levenshtein==0.12.2
thefuzz==0.19.0
bcdata>=0.4.5
